requirements.txt
streamlit==1.31.0
google-genai==0.2.2
pinecone>=5.0.0
pypdf2==3.0.1
python-dotenv==1.0.0
tiktoken==0.5.2

